coupon backend

+++ database/migrations/2022_04_30_142029_create_coupon_templates_table.php
@@ -18,7 +18,7 @@ class CreateCouponTemplatesTable extends Migration
             $table->string('title');
             $table->string('description')->nullable();
             $table->tinyInteger('type')->default(1);
-            $table->decimal('discount');
+            $table->integer('discount');
             $table->timestamps();
         });
     }
